Fixed worked hours in the trayicon for the first task of a day
The current day is updated on both sides of the record-toggle operation. Previously, the big number in the icon stayed from the previous day during the first task of the day, because current workday was only updated when recording stopped. Now, the workday is updated when the recording starts as well.